Add files via upload
This analysis focuses on understanding the performance of certain stores (trial stores), i.e [77, 86, 88], in a specific period known as trial period, which is from the start of february 2019 to the end of April 2019. In order to do this, certain stores which showed similar characteristics to the trial stores before the trial period were assigned as control to better understand the impact of the trial period. The characteristic observed was total sales revenue, total number of customers, as well as average number of transactions. After obtaining the control stores for each trial stores, a study was done to observe if total sales were higher in this period and, if so, understand if the driver for change is the total number of customers or the average transaction per customer. By so doing, appropriate measures can be taken to optimize the overall performance of each store